chart/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pydantic==2.7.4
pyswisseph==2.10.3.2
timezonefinder==6.5.2
python-dotenv==1.0.1
